main.rbに統合
